+++ views/options.php
@@ -1,0 +1,27 @@
+<div class="inner-meta">
+
+	<label>Subtitle</label>
+	<input type="text" name="subtitle" value="<?php echo $custom['subtitle'][0]; ?>" >
+	<br>
+
+	<label>Blurb</label>
+	<input type="text" name="blurb" value="<?php echo $custom['blurb'][0]; ?>" >
+	<br>
+
+	<label>Venue</label>
+	<input type="text" name="venue" value="<?php echo $custom['venue'][0]; ?>" >
+	<br>
+
+	<label>Venue URL</label>
+	<input type="text" name="venue" value="<?php echo $custom['venue_url'][0]; ?>" >
+	<br>
+
+	<label>Start Date</label>
+	<input class="prod_date" type="text" name="start_date" value="<?php echo $custom['start_date'][0]; ?>" size="15" placeholder="MM/DD/YYYY"/>
+	<br>
+
+	<label>End Date</label>
+	<input class="prod_date" type="text" name="start_date" value="<?php echo $custom['start_date'][0]; ?>" size="15" placeholder="MM/DD/YYYY"/>
+	<br>
+
+</div>